Getting references working
