+++ src/fetch-tools/index.js
@@ -1,6 +1,6 @@
-import fetchUrl from "./fetchUrl";
-import getRequestHeaders from "./getRequestHeaders";
-import selectUserAgent from "./selectUserAgent";
+import fetchUrl from "./fetchUrl.js";
+import getRequestHeaders from "./getRequestHeaders.js";
+import selectUserAgent from "./selectUserAgent.js";
 
 export default {
 	getRequestHeaders,
