+++ src/index.js
@@ -4,7 +4,7 @@ const createLinkArchive = require("./link-archiver");
 const uidLink = require("./link-uid");
 const createLinkHTMLCard = require("./link-block-maker");
 
-import setConfig from "./config";
+const setConfig = require("./config");
 
 const context = async (link, isArchiveLink, canonicalLink, timeout) => {
 	setConfig({ timeout: timeout || 35000 });
